parsingMultipartFile 🌸

+++ src/main/java/com/sprint/mission/discodeit/controller/UserController.java
@@ -1,19 +1,22 @@
 package com.sprint.mission.discodeit.controller;
 
-import com.sprint.mission.discodeit.common.Util;
 import com.sprint.mission.discodeit.entity.dto.*;
 import com.sprint.mission.discodeit.service.basic.UserService;
 import com.sprint.mission.discodeit.service.basic.UserStatusService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestPart;
+import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.IOException;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 import java.util.UUID;
 
+import static com.sprint.mission.discodeit.common.Util.parsingMultipartFile;
 import static org.springframework.web.bind.annotation.RequestMethod.*;
 
 @RestController
@@ -29,24 +32,12 @@ public class UserController {
 //    [ ] 모든 사용자를 조회할 수 있다.
 //    [ ] 사용자의 온라인 상태를 업데이트할 수 있다.
 
-
-
     //!! @Valid 검증 == dependencies 'spring-boot-starter-validation'
     @RequestMapping(value = "/create", method = POST)
     public Res_User create(@RequestPart("dtouser") Dto_User dtoUser,
                            @RequestPart("file") MultipartFile file) {
 
-        Dto_BinaryContent dtoFile = null;
-        try {
-            dtoFile = Dto_BinaryContent.from(
-                                                file.getOriginalFilename(),
-                                                file.getContentType(),
-                                                file.getBytes(),
-                                                file.getSize());
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-        Util.okMessage("file.getOriginalFilename() = [" + file.getOriginalFilename() + "]");
+        Dto_BinaryContent dtoFile = parsingMultipartFile(file);
         return userService.create(dtoUser, Optional.ofNullable(dtoFile));
     }
 
@@ -54,23 +45,15 @@ public class UserController {
     public Res_User update(@RequestPart("dtouser") Dto_UserWithIDAndContent dtoUser,
                            @RequestPart("file") MultipartFile file) {
 
-        Dto_BinaryContent dtoFile = null;
-        try {
-            dtoFile = Dto_BinaryContent.from(
-                    file.getOriginalFilename(),
-                    file.getContentType(),
-                    file.getBytes(),
-                    file.getSize());
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-        Util.okMessage("file.getOriginalFilename() = [" + file.getOriginalFilename() + "]");
-
+        Dto_BinaryContent dtoFile = parsingMultipartFile(file);
         return userService.update(dtoUser.userId(), dtoUser.dtoUser(), Optional.ofNullable(dtoFile));
     }
 
     @RequestMapping(value = "/delete/{id}", method = DELETE)
     public void delete(@PathVariable("id") UUID userID) {
+        if (userService.find(userID) == null) {
+            throw new NoSuchElementException("userController.delete.userId = [" + userID + "] not found");
+        }
         userService.delete(userID);
     }
